small tweaks and test fixes/cleanup
